+++ test/fixture/verification/called-with/never/supported.php
@@ -1,0 +1,5 @@
+<?php
+
+$message = 'Requires PHP >= 7.4';
+
+return version_compare(PHP_VERSION, '7.4.x', '>=');