+++ test/suite/Call/CallVerifierWithGeneratorsTest.php
@@ -40,9 +40,17 @@ class CallVerifierWithGeneratorsTest extends TestCase
         $this->returnedEvent = $this->callEventFactory->createReturned($this->returnValue);
         $this->call = $this->callFactory->create($this->calledEvent, $this->returnedEvent, null, $this->returnedEvent);
 
+        $this->arraySequencer = new Sequencer();
         $this->objectSequencer = new Sequencer();
         $this->invocableInspector = new InvocableInspector();
-        $this->exporter = new InlineExporter(1, $this->objectSequencer, $this->invocableInspector);
+        $this->featureDetector = FeatureDetector::instance();
+        $this->exporter = new InlineExporter(
+            1,
+            $this->arraySequencer,
+            $this->objectSequencer,
+            $this->invocableInspector,
+            $this->featureDetector
+        );
         $this->matcherFactory =
             new MatcherFactory(AnyMatcher::instance(), WildcardMatcher::instance(), $this->exporter);
         $this->matcherVerifier = new MatcherVerifier();
@@ -51,7 +59,6 @@ class CallVerifierWithGeneratorsTest extends TestCase
         $this->assertionRecorder = ExceptionAssertionRecorder::instance();
         $this->callVerifierFactory = CallVerifierFactory::instance();
         $this->assertionRecorder->setCallVerifierFactory($this->callVerifierFactory);
-        $this->featureDetector = FeatureDetector::instance();
         $this->differenceEngine = new DifferenceEngine($this->featureDetector);
         $this->differenceEngine->setUseColor(false);
         $this->assertionRenderer = new AssertionRenderer(