+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## Next release
 
 - **[BC BREAK]** PHP 7.1 is no longer supported.
+- **[IMPROVED]** Improved exporter output for array references ([#244]).
 - **[IMPROVED]** Improved exporter output for weak references ([#245]).
 
+[#244]: https://github.com/eloquent/phony/issues/244
 [#245]: https://github.com/eloquent/phony/issues/245
 
 ## 3.2.0 (2019-10-30)